refactor: organizing tests
